+++ src/LASRcore/PointCloud.h
@@ -24,6 +24,7 @@ public:
   bool add_attributes(const std::vector<Attribute>&);
   bool add_point(const Point& p);
   bool add_rgb();
+  bool remove_attribute(const std::string&);
   bool seek(size_t pos);
   bool read_point(bool include_withhelded = false);
   void set_file(const std::string& file) { this->file = file; };